Rename bfxapi.utils to _utils (and update references).
